now with IR remote control
The prior version required running a cable from the button to the Arduino Mega. This picked up major interference from the Magic Wand, causing the 7-segment LEDs to randomly blink and occasionally requiring a restart (even when using shielded cable). This commit solves the problem (🤞) by removing this cable. Now, the button press triggers an Adafruit Feather to blink IR LEDs, which are picked up by an IR receiver connected to the Arduino Mega.